+++ include/kinetra.h
@@ -6,7 +6,7 @@
 #include <string.h>
 #include <stdbool.h>
 
-#define KINETRA_VERSION "0.0.1a15"
+#define KINETRA_VERSION "0.0.1a16"
 #define MAX_TOKENS 4096
 #define MAX_AST_NODES 2048
 
@@ -78,6 +78,7 @@ typedef struct {
 	char lexeme[256];
 	double value;
 	int line;
+	int column;
 } Token;
 
 // --- AST Definitions ---